* commit 'e2c2a32ff5289eb6311e2b9129b08fe1d13dae97':
  KP-2388 Merge with develop: use Long
  KP-2388 Rename to FLOAT_CONSTANT to be in sync with VTL specification
  KP-2388 Cleanup: remove usages of Integer
  KP-2388 Cleanup: remove usages of Float and Integer
  KP-2388 Fix ConditionalExpressionVisitor: use only Long and Double as implementation for integer and float from VTL specification

+++ java-vtl-script/src/main/java/no/ssb/vtl/script/operations/check/CheckSingleRuleOperation.java
@@ -26,7 +26,7 @@ public class CheckSingleRuleOperation extends AbstractUnaryDatasetOperation {
     private final RowsToReturn rowsToReturn;
     private final ComponentsToReturn componentsToReturn;
     private final String errorCode;
-    private final Integer errorLevel;
+    private final Long errorLevel;
     private final Set<Component> conditions = Sets.newHashSet();
 
     private CheckSingleRuleOperation(Builder builder) {
@@ -167,7 +167,7 @@ public class CheckSingleRuleOperation extends AbstractUnaryDatasetOperation {
         builder.put(ERROR_CODE_LABEL, Component.Role.ATTRIBUTE, String.class);
 
         if (errorLevel != null) {
-            builder.put(ERROR_LEVEL_LABEL, Component.Role.ATTRIBUTE, Integer.class);
+            builder.put(ERROR_LEVEL_LABEL, Component.Role.ATTRIBUTE, Long.class);
         }
 
         return builder.build();
@@ -191,7 +191,7 @@ public class CheckSingleRuleOperation extends AbstractUnaryDatasetOperation {
         private RowsToReturn rowsToReturn = RowsToReturn.NOT_VALID;
         private ComponentsToReturn componentsToReturn = ComponentsToReturn.MEASURES;
         private String errorCode;
-        private Integer errorLevel;
+        private Long errorLevel;
 
         public Builder(Dataset dataset) {
             this.dataset = dataset;
@@ -212,7 +212,7 @@ public class CheckSingleRuleOperation extends AbstractUnaryDatasetOperation {
             return this;
         }
 
-        public Builder errorLevel(Integer errorLevel) {
+        public Builder errorLevel(Long errorLevel) {
             this.errorLevel = errorLevel;
             return this;
         }
